requirements.txt
future==0.18.2
Pillow==9.1.1